Attempt to implement StreamSequenceTokens
Looks like PullingAgent has a lot of control, CacheCursor seems to be thrown away a lot.
Have to have a link about how best to intercept this, the control message at first is buggering up initial subscription.
Can we perform a call to GetCacheCursor *before* the we fire a control message?